rename addItem to setQuantity
