+++ src/intentDelegates/incrementQuantity.js
@@ -14,7 +14,7 @@ module.exports = function(handler, table) {
                 })
         .then(function(resp) {
             if (resp == undefined) {
-                handler.emit('AddItem')
+                handler.emit('SetQuantity')
             } else {
               if (resp.quantity != undefined) {
                 baseQuantity = resp.quantity;
